requirements.txt
Flask==0.10.1
Flask-MySQL==1.4.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pkg-resources==0.0.0
PyMySQL==0.8.1
Werkzeug==0.14.1
